Improve detection & debugging of friends' changes.
Modify the detection of changes to friends' details so that
modifications to the number of items on their wall are
ignored. Longer term, we should probably look at the changes
themselves, ignoring entries that expire and notifying the
user of new entries.

In addition, modify the debugging code so that we only print
the old and new values where there has been a change.

Signed-off-by: Nigel Cunningham <nigel@tuxonice.net>